+++ Rdmp.UI/CommandExecution/AtomicCommands/ExecuteCommandViewParentTree.cs
@@ -1,11 +1,16 @@
-﻿using Rdmp.Core.Curation.Data;
+﻿// Copyright (c) The University of Dundee 2024-2024
+// This file is part of the Research Data Management Platform (RDMP).
+// RDMP is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+// RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+// You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
+
 using Rdmp.Core.Providers;
 using Rdmp.UI.SimpleDialogs;
+using SixLabors.ImageSharp;
+using SixLabors.ImageSharp.PixelFormats;
 using System;
-using System.Collections.Generic;
+using System.Drawing;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Rdmp.Core.CommandExecution.AtomicCommands;
 
@@ -20,8 +25,6 @@ public class ExecuteCommandViewParentTree : BasicCommandExecution, IAtomicComman
     {
         _activator = activator;
         _databaseObject = databaseObject;
-        //var type = databaseObject.GetType();
-        //Console.WriteLine(type.Name);
     }
 
     private void BuildTree()